*net-xml-generator-version* 1.0.2
If an element attibute value is nil, skip generating that attribute
value pair.

Escape any #\> in xml-write, in an attribute value, in a value after
an #\@ in element content, and in a string at top level in element
content.  All this is to obey the restriction that the #\> _must_ be
escaped anywhere it appears in an XML document other than as the close
marker of a CDATA section.