net-xml-generator version 1.1.1
Version from 1.0.2 => 1.1.1

The #\@ syntactic marker can now be doubled as described in release
notes.  Added an example in the blurb.

Cleaned up some inline documentation, which is the only documentation
other than the xml-generator-blurb example.

Suppressed more noise files in .gitignore.

<release_notes>
net-xml-generator version 1.1.1

The @@ synrax marker can now be doubled, suppressing XML escaping of
an element content string.  This allows an application to insert
preformatted text without interference from the XML generator.

Doubling an @@ which marks an attribute value inside a start tag
causes that attval not to be enclosed in double quotes.  This syntax
seems nexessary in some JavaScript contexts, e.g.:

    ^((div @@foo.onLoad "init())" => <div foo.onLoad=init()>

The changes in this version are back compatible.

</release_notes>

Change-Id: Ieb3e0a7e46ee93bb9409762b1c80ab84baf34937
Reviewed-on: https://gerrit.franz.com:9080/5964
Reviewed-by: Kevin Layer <layer@franz.com>
Tested-by: Kevin Layer <layer@franz.com>